requirements.txt
fastapi
uvicorn
requests
numpy
Pillow
